core: v2 polish (#107)
* chore: updates

* updates

* updates

* updates

* updates

* updates

* 2d working

* updates

* updates

* updates

* working webgl

* updates

* updates

* updates

* updates

* Update Cargo.toml

* updates

* updates

* updates

* updates

* Update webgl.rs

* updates

* updates

* Update .gitignore

* updates

* updates

* updates

* working much better :D

* updates

* updates

* chore: refactor

* chore: clean up

* chore: clean up

* chore: clean up

* feat: JSI

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: clean up

* chore: updates

* chore: clean up

* chore: clean up

* chore: clean up

* chore: updates

* chore: clean up

* chore: clean up

* chore: updates

* chore: updates

* chore: updates

* Update WebGL2RenderingContext.cpp

* chore: use namespace

* chore: updates

* chore: updates

* chore: fix build

* chore: initial 2d support

* fix: gl alive!!!

* fix: updates

* fix: working 3D

* fix: 2d

* fix: video rendering

* fix: 2d context to 2d context drawing

* feat: v2 alpha.0

* chore: updates & trying another op

* fix: remove extra conversion

* chore: clean up

* chore: updates

* chore: updates

* chore: bump

* feat: initial dom view

* feat: expose skia dom

* chore: ios updates (#99)

* chore: ios updates

chore: replaces Arc with Rc

* fix(android): updates

* chore: ios updates

* chore: ios updates

* chore: ios updates

* chore: updates

* chore: updates

* chore: updates

* fix: ios

* chore: updates

* chore: updates

* fix(ios): missing pods

* fix(ios): build

* chore: updates

* chore: updates

* chore: updates

* fix(ios): improve fileop

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* Chore v2 perf op (#104)

* chore: updates

* chore: updates

* chore: nx migrate @nativescript/plugin-tools

* release: 2.0.0-alpha.17

* fix: fenceSync

* fix(jsi): fenceSync

* chore: updates

* chore: upgrade skia

* feat: v8 refactor (#103)

* feat: v8 refactor

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: improve code sharing

* chore: updates

* chore: updates

* chore: updates

* chore: updates

* chore: working ios

* chore: update

* fix: running on both platforms

* chore: updates

* chore: 2.0.0-alpha.21

* feat:  the frantic run of the valorous rabbit demo game

* chore: alpha 22

* chore: updates

* chore: updates

* chore: updates

* chore: update

* fix: package building

* chore: updates

* chore: updates

---------


* chore: updates

* chore(android): perf op

* chore: updates

* fix: font handling & custom font support

* chore: updates

* Squashed commit of the following:

commit ec98277b3ae368574032e3ab5f206feb76462b5c
Author: Dylan Llewellyn <46717769+herefishyfish@users.noreply.github.com>
Date:   Fri Dec 8 03:47:51 2023 +0800

    fix: add text baseline bindings (#106)

    * fix: add text baseline bindings

    * fix: add text baseline bindings

    * fix: add text baseline bindings

commit b355ddc66cbbe1edff8832353e2dad2535f73de5
Author: Osei Fortune <fortune.osei@yahoo.com>
Date:   Thu Nov 23 07:00:15 2023 -0400

    chore: updates (#105)

* fix: baseline middle

* chore: updates

* chore: updates

* chore: bump

* fix: scaling

* fix: handle empty text when drawing

* chore: updates

---------

+++ packages/canvas/src-native/canvas-native/canvas-2d/src/image_bitmap.rs
@@ -249,7 +249,7 @@ pub(crate) fn create_image_bitmap_internal(
         Some(mut surface) => {
             let canvas = surface.canvas();
             if flip_y {
-                canvas.translate(skia_safe::Vector::new(0., source_rect.height() as f32));
+                canvas.translate(skia_safe::Vector::new(0., source_rect.height()));
                 canvas.scale((1., -1.));
             }
